Revert "issue tests (#1)"
This reverts commit 03751c31797c52525c4cf93e511c2d59b5692cc2.
