+++ .eslintrc.js
@@ -3,7 +3,6 @@ module.exports = {
     commonjs: true,
     es6: true,
     node: true,
-    mocha: true,
   },
   extends: [
     'airbnb-base',
