+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "start": "node src/index.js",
     "dev": "nodemon src/index.js",
-    "lint": "eslint src/",
-    "test": "nodemon --exec \"mocha src/*.test.js src/**/*.test.js\""
+    "lint": "eslint src/"
   },
   "keywords": [],
   "author": "",
@@ -21,12 +20,9 @@
     "morgan": "^1.9.1"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.20.1",
-    "mocha": "^7.1.0",
-    "nodemon": "^2.0.2",
-    "supertest": "^4.0.2"
+    "nodemon": "^2.0.2"
   }
 }
