+++ src/api/issues.js
@@ -1,7 +1,6 @@
 const { Router } = require('express');
 
 const IssueEntry = require('../models/issueEntry');
-const middleware = require('./issues.middlewares');
 
 const router = Router();
 
@@ -14,25 +13,27 @@ router.get('/', async (req, res, next) => {
   }
 });
 
-router.post('/', middleware.checkProperties, middleware.validateIssue, async (req, res, next) => {
+router.post('/', async (req, res, next) => {
   try {
     const issueEntry = new IssueEntry(req.body);
     const createdEntry = await issueEntry.save();
     res.json(createdEntry);
   } catch (error) {
+    if (error.name === 'ValidationError') {
+      res.status(422);
+    }
     next(error);
   }
 });
 
 
-router.get('/:id', middleware.isValidIdFormat, async (req, res, next) => {
+router.get('/:id', async (req, res, next) => {
   const issueId = req.params.id;
   try {
-    const entry = await IssueEntry.findOne({ _id: issueId });
-    if (entry) {
+    if (issueId) {
+      const entry = await IssueEntry.findOne({ _id: issueId });
       res.json(entry);
     } else {
-      res.status(404);
       throw new Error('Invalid issue ID submitted');
     }
   } catch (error) {
@@ -40,34 +41,16 @@ router.get('/:id', middleware.isValidIdFormat, async (req, res, next) => {
   }
 });
 
-router.patch(
-  '/:id',
-  middleware.isValidIdFormat,
-  middleware.checkProperties,
-  async (req, res, next) => {
-    const issueId = req.params.id;
-    try {
-      await IssueEntry.findByIdAndUpdate(issueId, req.body, { new: true }, (err, issue) => {
-        if (!err) { res.status(200).json(issue); }
-      }).catch((error) => {
-        res.status(422);
-        throw new Error('Can not update issue');
-      });
-    } catch (error) {
-      next(error);
-    }
-  },
-);
-
-router.delete('/:id', middleware.isValidIdFormat, async (req, res, next) => {
+router.patch('/:id', async (req, res, next) => {
   const issueId = req.params.id;
   try {
-    const entry = await IssueEntry.findOne({ _id: issueId });
-    if (entry) {
-      entry.remove();
-    }
-    res.status(204);
-    res.json();
+    // TODO validate req.body against schema
+    await IssueEntry.findByIdAndUpdate(issueId, req.body, { new: true }, (err, issue) => {
+      if (!err) { res.status(200).json(issue); }
+    }).catch((error) => {
+      res.status(422);
+      throw new Error('Can not update issue', error);
+    });
   } catch (error) {
     next(error);
   }
