+++ src/index.js
@@ -1,4 +1,39 @@
-const app = require('./app');
+const express = require('express');
+const morgan = require('morgan');
+const helmet = require('helmet');
+const cors = require('cors');
+const mongoose = require('mongoose');
+
+require('dotenv').config();
+
+const middlewares = require('./middlewares');
+const issues = require('./api/issues');
+
+const app = express();
+
+mongoose.connect(process.env.DATABASE_URL, {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  useFindAndModify: false,
+});
+
+app.use(morgan('common')); // logs requests
+app.use(helmet()); // blocks some headers and adds others
+app.use(cors({
+  origin: process.env.CORS_ORIGIN,
+})); // needed to allow other sites to use server
+app.use(express.json());
+
+app.get('/', (req, res) => {
+  res.json({
+    message: 'Hello World!',
+  });
+});
+
+app.use('/api/v1/issues', issues);
+
+app.use(middlewares.notFound);
+app.use(middlewares.errorHandler);
 
 const port = process.env.PORT || 3000;
 app.listen(port, () => {
