+++ src/models/issueEntry.js
@@ -1,31 +1,18 @@
 const mongoose = require('mongoose');
 
-// eslint-disable-next-line no-unused-vars
-const db = require('./../db/connection');
-
 const { Schema } = mongoose;
 
+const requiredString = {
+  type: String,
+  required: true,
+};
+
 const issueEntrySchema = new Schema({
-  project: {
-    type: String,
-    required: [true, '"project" is required'],
-  },
-  type: {
-    type: String,
-    required: [true, '"type" is required'],
-  },
-  status: {
-    type: String,
-    required: [true, '"status" is required'],
-  },
-  priority: {
-    type: String,
-    required: [true, '"priority" is required'],
-  },
-  summary: {
-    type: String,
-    required: [true, '"summary" is required'],
-  },
+  project: requiredString,
+  type: requiredString,
+  status: requiredString,
+  priority: requiredString,
+  summary: requiredString,
   description: String,
 }, {
   timestamps: true,
